striimu-services/backend/requirements.txt
Flask
Flask-Bcrypt
Flask-Cors
Flask-WTF
JWT
peewee
python-dotenv
bcrypt
python-jose
password-validator
